♻️refactor: Search page style refactoring
