+++ src/components/BookSearchPage/BestSellerSection.tsx
@@ -1,0 +1,20 @@
+import { Box, Typography } from '@mui/material';
+import BestBookCarousel from './BestBookCarousel';
+
+const BestSellerSection = (): JSX.Element => {
+  return (
+    <Box
+      sx={{
+        width: '100%',
+        marginBottom: 5,
+      }}
+    >
+      <Typography variant="h3" fontWeight="bold" gutterBottom>
+        베스트셀러
+      </Typography>
+      <BestBookCarousel />
+    </Box>
+  );
+};
+
+export default BestSellerSection;
