+++ src/components/BookSearchPage/BookSearchBar.tsx
@@ -1,5 +1,6 @@
-import { TextField, InputAdornment } from '@mui/material';
+import { TextField, InputAdornment, Box } from '@mui/material';
 import SearchIcon from '@mui/icons-material/Search';
+import { bookSearchBarStyles } from './BookSearch.style';
 
 interface BookSearchBarProps {
   value: string;
@@ -19,29 +20,30 @@ const BookSearchBar = ({
   };
 
   return (
-    <TextField
-      variant="outlined"
-      placeholder="도서명, 저자를 검색하세요"
-      fullWidth
-      value={value}
-      onChange={onChange}
-      onKeyDown={handleKeyDown}
-      InputProps={{
-        startAdornment: (
-          <InputAdornment
-            sx={{ cursor: 'pointer' }}
-            onClick={onSearch}
-            position="start"
-          >
-            <SearchIcon />
-          </InputAdornment>
-        ),
-      }}
-      sx={{
-        width: '100%',
-        backgroundColor: '#fff',
-      }}
-    />
+    <Box sx={bookSearchBarStyles.bookSearchBarBox}>
+      <TextField
+        variant="outlined"
+        placeholder="도서명, 저자를 검색하세요"
+        fullWidth
+        value={value}
+        onChange={onChange}
+        onKeyDown={handleKeyDown}
+        slotProps={{
+          input: {
+            startAdornment: (
+              <InputAdornment
+                sx={{ cursor: 'pointer' }}
+                onClick={onSearch}
+                position="start"
+              >
+                <SearchIcon />
+              </InputAdornment>
+            ),
+          },
+        }}
+        sx={bookSearchBarStyles.textfield}
+      />
+    </Box>
   );
 };
